e2fsck: Dump message buffer
From: Amir Goldstein <amir73il@users.sf.net>

Error messages are recorded in a message buffer after the
journal super block.  On journal recovery, the journal message buffer
is copied to the file system message buffer.  On fsck, if the message
buffer is not empty, the recorded messages are printed to stdout and
the buffer is cleared.
With a default block size of 4K, there is always 2K of free
space for the message buffer after the 1K super block.

Signed-off-by: Amir Goldstein <amir73il@users.sf.net>